Add menu item to explain why some controls are disabled
While the system menu is not the best place to put this,
the app's UI layout precludes (easily) placing this command
elsewhere.

I thought that a new user might be wondering why the
Auto Update settings are always disabled in the Store version,
even if the application has been launched as an administrator.